Fix printf format warnings in B_TRANSLATE strings
Previously these were not checked by the compiler due to a missing
function attribute.

Unfortunately, for the translation macros to work, the translated string
must be identical in all architectures (otherwise, we would need
different translations and different catalogs for each). This means the
B_PRI* macros can't be used, and instead the parameters must be in one
of the types handled directly by printf. Change the variables types
directly where it was easily possible, and otherwise, use a cast.

+++ src/add-ons/translators/ico/ICOTranslator.cpp
@@ -119,8 +119,8 @@ ICOTranslator::DerivedIdentify(BPositionIO *stream,
 	info->capability = ICO_IN_CAPABILITY;
 	snprintf(info->name, sizeof(info->name), 
 		B_TRANSLATE("Windows %s %ld bit image"),
-		type == ICO::kTypeIcon ? B_TRANSLATE("Icon") : B_TRANSLATE("Cursor"), 
-		bitsPerPixel);
+		type == ICO::kTypeIcon ? B_TRANSLATE("Icon") : B_TRANSLATE("Cursor"),
+		(long int)bitsPerPixel);
 	strcpy(info->MIME, kICOMimeType);
 
 	return B_OK;
